fix: detach eeg analysis
- /eeg/analysis 내에 있는 검사 지표들을 각 파일로 분리
- brain_connectivity, brain_connectivity_diff는 coherence, phase lock value 두가지 method를 사용하는데, 해당 코드를 통합시킴.
