+++ utils/eeg/eeg_analysis/brain_connectivity.py
@@ -1,0 +1,79 @@
+import copy
+import mne
+from mne_connectivity import spectral_connectivity_epochs
+from mne_connectivity.viz import plot_connectivity_circle
+import matplotlib.pyplot as plt
+import os
+import cv2
+import base64
+
+def center_crop(img, dim):
+    width, height = img.shape[1], img.shape[0]
+
+    # process crop width and height for max available dimension
+    crop_width = dim[0] if dim[0] < img.shape[1] else img.shape[1]
+    crop_height = dim[1] if dim[1] < img.shape[0] else img.shape[0]
+    mid_x, mid_y = int(width / 2), int(height / 2)
+    cw2, ch2 = int(crop_width / 2), int(crop_height / 2)
+    crop_img = img[mid_y-ch2:mid_y+ch2, mid_x-cw2:mid_x+cw2]
+    return crop_img
+
+def get_brain_connectivity(epoch_data, uuid, type):
+    epoch_data = copy.deepcopy(epoch_data)
+    eeg_info = epoch_data.info
+    sfreq = eeg_info['sfreq']
+    epoch_data = epoch_data.get_data()
+    sample_size = epoch_data.shape[0]
+    step = sample_size // 5
+    exp_names = ['baseline', 'stimulation1', 'recovery1', 'stimulation2', 'recovery2']
+    bands = {'delta': [0, 4], 'theta': [4, 8], 'alpha': [8, 13], 'beta': [13, 30], 'gamma': [30, 40]}
+    files = {exp_name: {band_name: None for band_name in bands.keys()} for exp_name in exp_names}
+
+    fc_method = type
+    if(type == 'coh'):
+        dir = 'connectivity'
+    elif(type == 'plv'):
+        dir = 'connectivity2'
+
+    for i in range(5):
+        start, end = i * step, (i+1) * step
+        sample = epoch_data[start: end, ...]
+        raw = mne.EpochsArray(sample, info=eeg_info)
+
+        for band_name, band_range in bands.items():   
+
+            start, end = i * step, (i+1) * step
+            sample = epoch_data[start: end, ...]
+            raw = mne.EpochsArray(sample, info=eeg_info)
+            con = spectral_connectivity_epochs(data=raw, names=raw.info['ch_names'], sfreq=sfreq,
+                                               mode="multitaper",
+                                               method=fc_method, fmin=band_range[0], fmax=band_range[1],
+                                               faverage=True,
+                                               mt_adaptive=True, n_jobs=1)
+            conmat = con.get_data(output="dense")[:, :, 0]
+            tmp_name = os.path.join('image', dir, '{}_{}_{}.jpg'.format(uuid, exp_names[i], band_name))
+            tmp_name = os.path.abspath(tmp_name)
+
+            fig, ax = plt.subplots(1, figsize=(5, 5), subplot_kw={'projection': 'polar'})
+            plot_connectivity_circle(conmat, raw.info['ch_names'], n_lines=50,
+                                     facecolor='white', textcolor='black',
+                                     colormap='gray_r',
+                                     ax=ax,
+                                     colorbar=False,
+                                     show=False,
+                                     node_linewidth=1)
+            fig.savefig(tmp_name)
+            image = cv2.imread(tmp_name)
+            crop_img = center_crop(image, (370, 370))
+            cv2.imwrite(tmp_name, crop_img)
+
+            plt.close('all')
+            plt.clf()
+
+            with open(tmp_name, 'rb') as f:
+                im_bytes = f.read()
+            im_b64 = base64.b64encode(im_bytes).decode("utf8")
+            files[exp_names[i]][band_name] = im_b64
+    plt.close('all')
+    plt.clf()
+    return files
